// client/src/core/types/form.d.ts
export type ContractFilterForm = {
  school: {
    id: number
    schoolName: string
  } | null
  classCustomers: classCustomer[]
  privateCustomers: privateCustomer[]
  subject: subject | null
  interval: number
  minStartDate: Dayjs | null
  startDate: Dayjs | null
  endDate: Dayjs | null
  dow: number | undefined
  startTime: Dayjs | null
  endTime: Dayjs | null
  customerType: CustomerType
  contractType: ContractType
}
export type ContractFilterFormErrorTexts = {
  school: string
  classCustomers: string
  privateCustomers: string
  subject: string
  interval: string
  startDate: string
  endDate: string
  customerType: string
  contractType: string
}

export type ContractCreationForm = {
  startDate: Dayjs | null
  endDate: Dayjs | null
  startTime: Dayjs | null
  endTime: Dayjs | null
  minTime: Dayjs | null
  maxTime: Dayjs | null
  teacher: string
  teacherConfirmation: boolean
  dow: number
  selsuggestion: string
}